Start of refactor
- Break up `header.rs` into smaller files.
- Move header-related parsing into `header.rs`
- Rename the cryptic `RCode` to the much more sensible `ReturnCode`.
